Add images and QFA data for stripes
Tried to add all images to repo, but was disaster.  Too much data for
github.  Noticed that .git directory was 7Gb!!  Wondered if that was
because of my aborted attempt to commit files?  Searched commit history
for files >10MB.  Only found a handful.  Ran git gc.  Reduced to 900MB,
but still don't understand why so large.